work in progress..
* refactor the 'voprf' crypto module into a reusable class,
  so each provider can own a distinct instance

* add the dependency: 'keccak'
  - v2 of the extension included a snapshot of this library
  - v3 removed it, but there is still code that wants to use it
    * presumably, this branch of code will get called
      when the "hCaptcha" provider is reimplemented (..coming soon)
