+++ package.json
@@ -1,11 +1,13 @@
 {
     "name": "privacy-pass",
-    "version": "3.0.0",
+    "version": "3.2.0",
+    "private": true,
     "contributors": [
         "Suphanat Chunhapanya <pop@cloudflare.com>",
-        "Armando Faz <armfazh@cloudflare.com>"
+        "Armando Faz <armfazh@cloudflare.com>",
+        "Warren Bank <warren.r.bank@gmail.com>"
     ],
-    "main": "index.js",
+    "browser": "index.js",
     "license": "BSD-3-Clause",
     "type": "module",
     "scripts": {
@@ -21,7 +23,7 @@
         "asn1-parser": "^1.1.8",
         "axios": "^0.23.0",
         "buffer": "^6.0.3",
-        "keccak": "^3.0.1",
+        "keccak": "^3.0.2",
         "qs": "^6.10.1",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
@@ -33,6 +35,7 @@
     "devDependencies": {
         "@types/chrome": "^0.0.159",
         "@types/jest": "^27.0.2",
+        "@types/keccak": "^3.0.1",
         "@types/qs": "^6.9.6",
         "@types/react": "^17.0.5",
         "@types/react-dom": "^17.0.5",
