+++ public/manifest.json
@@ -1,7 +1,7 @@
 {
     "name": "__MSG_appName__",
     "description": "__MSG_appDescription__",
-    "version": "3.1.0",
+    "version": "3.2.0",
     "manifest_version": 2,
     "default_locale": "en",
     "icons": {
