+++ src/background/crypto/pseudorandom-number-generators/shake.ts
@@ -1,0 +1,6 @@
+import createKeccakHash, { Shake, ShakeAlgorithm } from 'keccak';
+
+export function createShake256(): Shake {
+    const algorithm: ShakeAlgorithm = 'shake256';
+    return createKeccakHash(algorithm);
+}
