+++ src/background/crypto/voprf.test.ts
@@ -1,0 +1,8 @@
+import { VOPRF, defaultECSettings } from './voprf';
+
+test('randomPoint', () => {
+    const voprf = new VOPRF(defaultECSettings);
+    const P     = voprf.newRandomPoint();
+
+    expect(P).toBeDefined();
+});
