--- src/background/voprf.test.ts
@@ -1,7 +1,0 @@
-import { defaultECSettings, initECSettings, newRandomPoint } from './voprf';
-
-test('randomPoint', () => {
-    initECSettings(defaultECSettings);
-    const P = newRandomPoint();
-    expect(P).toBeDefined();
-});
